backend/pyproject.toml
[project]
name = "langchain-practice"
version = "0.1.0"
description = "LangChain + LangGraph Chatbot"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core backend
    "fastapi",
    "uvicorn",
    "pydantic",
    "requests",

    # LangChain ecosystem
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-google-genai",
    "langgraph",

    # Utilities
    "python-dotenv",
    "tenacity",
    "PyYAML",
    "jsonpatch",
    "dataclasses-json",

    # Optional / future use
    "numpy",
    "SQLAlchemy",
    "langsmith",

    # Frontend
    "streamlit",
]

[tool.uvicorn]
app = "app.main:app"
host = "0.0.0.0"
port = 8000
reload = true
